RCPS-84 Student Information Management: Fix Student model
