+++ src/main/java/com/student/information/management/app/model/student/Student.java
@@ -113,112 +113,5 @@ public class Student {
     /**
      * Validation for student id should be have a pattern like CT00-0000.
      */
-    private static List<String> existingStudentIds = new ArrayList<>();
 
-    public static boolean isValidStudentId(String studentId) {
-        String pattern = "CT\\d{2}-\\d{4}";
-        if (studentId.matches(pattern)) {
-            if (existingStudentIds.contains(studentId)) {
-                System.out.println("Student ID " + studentId + " already exists.");
-                return false;
-            } else {
-                existingStudentIds.add(studentId);
-                return true;
-            }
-        } else {
-            System.out.println("Invalid student ID format.");
-            return false;
-        }
-    }
-
-    /**
-     * Validation for lastName it should be a string only.
-     */
-    public static boolean isValidLastName(String lastName) {
-        return !lastName.isEmpty() && lastName.matches("[a-zA-Z ]+");
-    }
-
-    /**
-     * Validation for firstName it should be a string only.
-     */
-    public static boolean isValidFirstName(String firstName) {
-        return !firstName.isEmpty() && firstName.matches("[A-a-zZ ]+");
-    }
-
-    /**
-     * Validation for middleName it should be a string only.
-     */
-    public static boolean isValidMiddleName(String middleName) {
-        return !middleName.isEmpty() && middleName.matches("[A-a-zZ ]+");
-    }
-
-    /**
-     * Validation for sex it should be a male or female only.
-     */
-    public static boolean isValidSex(String sex) {
-        return sex.equalsIgnoreCase("Male") || sex.equalsIgnoreCase("Female");
-    }
-
-    /**
-     * Validation for Birthday it should have this pattern MM/DD/YYYY.
-     */
-    public static boolean isValidBirthday(String birthday) {
-        String[] parts = birthday.split("/");
-        if (parts.length != 3)
-            return false;
-
-        int month, day, year;
-        try {
-            month = Integer.parseInt(parts[0]);
-            day = Integer.parseInt(parts[1]);
-            year = Integer.parseInt(parts[2]);
-        } catch (NumberFormatException e) {
-            return false;
-        }
-
-        if (month < 1 || month > 12)
-            return false;
-
-        if (day < 1 || day > 31)
-            return false;
-
-        if (year < 1950 || year > 2100)
-            return false;
-
-        return true;
-    }
-
-    /**
-     * Validation for religion it should be a string only.
-     */
-    public static boolean isValidReligion(String religion) {
-        return !religion.isEmpty() && religion.matches("[a-zA-Z ]+");
-    }
-
-    /**
-     * Validation for email it should have a @ and . input.
-     */
-    public static boolean isValidEmail(String email) {
-        String emailRegex = "^[a-zA-Z0-9_+&*-]+(?:\\.[a-zA-Z0-9_+&*-]+)*@(?:[a-zA-Z0-9-]+\\.)+[a-zA-Z]{2,7}$";
-        Pattern pattern = Pattern.compile(emailRegex);
-        Matcher matcher = pattern.matcher(email);
-        return matcher.matches();
-    }
-
-    /**
-     * Validation for address it should not be null.
-     */
-    public static boolean isValidAddress(String address) {
-        return address != null && !address.isEmpty();
-    }
-
-    /**
-     * Validation for contact number it should be a number only.
-     */
-    public static boolean isValidContactNumber(String contactNumber) {
-        String contactNumberRegex = "^[0-9]+$";
-        Pattern pattern = Pattern.compile(contactNumberRegex);
-        Matcher matcher = pattern.matcher(contactNumber);
-        return matcher.matches();
-    }
 }
